requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
chromedriver==2.24.1
chromedriver-binary-auto==0.1.1
click==8.0.1
fastapi==0.65.2
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
idna==2.10
jedi==0.18.0
lxml==4.6.3
msgpack==1.0.2
mypy==0.910
mypy-extensions==0.4.3
neovim==0.3.1
parso==0.8.2
pydantic==1.8.2
pynvim==0.4.3
requests==2.25.1
selenium==3.141.0
soupsieve==2.2.1
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
